+++ CHANGELOG.md
@@ -1,5 +1,8 @@
+v0.4.58 (unreleased)
+* Added -filter-geom command with bbox= option, for removing non-intersecting geometry.
+
 v0.4.57
-* Bug fix for issue #228 "keep_shapes option doesn't always keep shapes"
+* Bug fix for issue #228 "keep_shapes option doesn't always keep shapes".
 
 v0.4.56
 * -clean now removes empty geometries by default.